+++ src/main/java/com/supportsystem/application/config/SecurityConfig.java
@@ -1,7 +1,7 @@
 package com.supportsystem.application.config;
 
 import com.supportsystem.application.security.JwtFilter;
-import com.supportsystem.application.services.JwtUserDetailsService;
+import com.supportsystem.application.services.AppUserService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
@@ -14,6 +14,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
@@ -27,11 +28,11 @@ public class SecurityConfig {
 
     private final JwtFilter jwtFilter;
 
-    private final JwtUserDetailsService jwtUserDetailsService;
+    private final AppUserService userService;
 
     @Autowired
-    public SecurityConfig(JwtUserDetailsService jwtUserDetailsService, JwtFilter jwtFilter) {
-        this.jwtUserDetailsService = jwtUserDetailsService;
+    public SecurityConfig(AppUserService userService, JwtFilter jwtFilter) {
+        this.userService = userService;
         this.jwtFilter = jwtFilter;
     }
 
@@ -63,7 +64,7 @@ public class SecurityConfig {
     public DaoAuthenticationProvider daoAuthenticationProvider() {
         log.info("DaoAuthenticationProvider initialized");
         DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
-        provider.setUserDetailsService(jwtUserDetailsService);
+        provider.setUserDetailsService((UserDetailsService) userService);
         provider.setPasswordEncoder(passwordEncoder());
         return provider;
     }
