+++ src/main/java/com/supportsystem/application/security/JwtTokenUtil.java
@@ -22,7 +22,7 @@ public class JwtTokenUtil {
     }
 
     public Claims extractAllClaims(String token) {
-        return Jwts.parserBuilder()
+        return Jwts.parser()
             .setSigningKey(key)
             .build()
             .parseClaimsJws(token)
