+++ src/main/java/com/supportsystem/application/services/UserRegistrationService.java
@@ -1,0 +1,8 @@
+package com.supportsystem.application.services;
+
+import com.supportsystem.application.domains.AppUser;
+
+public interface UserRegistrationService {
+
+    AppUser registerUser(String username, String password, String email);
+}
